server: Update the BlockEventsProcessor flow
When a block is processed, in case it can't be fully retrieved from the
rpc server, the block is ignored.

When a block is processed, if there is an existing block having the same
height, it will be replaced by the new one.

This is a part for fixing the bug #6