server: Initial schema migration from hex strings to bytes
The HEX_STRING_TYPE is deleted and BYTEA is used instead, which
only stores the bytes instead of the hex-string representation.

The merkle_root column is changed from TEXT to BYTEA.

The transaction_outputs.hex_script column is changed from TEXT to BYTEA.

The disk requirements for these fields is reduced by half.