web-ui: Trezor Enhancements
Improve the tests, also, avoid spending TPoS contract collaterals, they need to
be manually cancelled.

Also, add a note about the TPoS integration working only for Trezor T, explaining that
the contract is only created, and the coins to stake need to be sent to the staking address
after that, which leads to remove the amount input.